requirements.txt
Django~=5.2.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-filter==25.1
django-cors-headers==4.3.1
django-tables2>=2.7.5
django-simpleui>=2025.1.13
django-widget-tweaks==1.5.0
python-ulid-django==1.2.1
django-q2==1.8.0
pydantic==2.11.5
python-dotenv==1.1.1
drf-spectacular==0.29.0
colorlog==6.9.0
#mysqlclient>=2.2.7
pymysql>=1.1.0
pycryptodome>=3.20.0
httpx>=0.27.0
langchain==1.2.13
langchain-core==1.2.21
langchain-openai==1.1.12
tiktoken>=0.12.0
langchain-anthropic==1.4.0
langchain-mcp-adapters==0.2.2
langgraph==1.1.3
langgraph-prebuilt==1.0.8  # 1.0.9 与 langgraph 1.1.3 不兼容（缺少 ServerInfo）
langgraph-checkpoint-redis==0.4.0
deepagents==0.4.12
langchain-dashscope>=0.1.6